pyproject.toml
[tool.poetry]
name = "bots"
version = "0.1.0"
description = ""
authors = ["Lu1z-Gust4v0 <gustavo1204.gc@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.20.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
